(GH-71) Removed manual separator property on files argument
By setting this property to a space character
this introduced problems if the paths of the filse
contained spaces. The commandline parser library
we use already handles spaces in file paths
without this and allows for multiple files paths
to be specified with a space between them
as long as each file path is correctly double
quoted (must be double quoted to handle spaces).

fixes #71