(GH-146) Add warning when git executable was not found
When the current repository is a git repository,
we should output a warning when there is no available
git executable available on the PATH environment variable
that we can use. This is not something that should stop
the execution of codecov-exe, as there are fallbacks
that can be used on CI systems.

fixes #146